requirements.txt
streamlit>=1.33
pandas>=2.2
openpyxl>=3.1
xlsxwriter>=3.2
python-dotenv>=1.0
openai>=1.30
